+++ admin-ui/app/utils/LicenseScreens/ApiKey.js
@@ -8,18 +8,7 @@ function ApiKey() {
   const dispatch = useDispatch()
   const serverError = useSelector((state) => state.licenseReducer.error)
   const isLoading = useSelector((state) => state.licenseReducer.isLoading)
-  const isLicenceAPIkeyValid = useSelector(
-    (state) => state.licenseReducer.isLicenceAPIkeyValid,
-  )
-  const params = {
-    apiKey: '',
-    productCode: '',
-    sharedKey: '',
-    managementKey: '',
-  }
 
-  const [values, setValues] = useState(params)
-  const [error, setError] = useState(params)
   const [submitted, setIsSubmitted] = useState(false)
   const [licenseKey, setLicenseKey] = useState('')
 
@@ -30,39 +19,6 @@ function ApiKey() {
       [name]: value,
     }))
   }
-
-  useEffect(() => {
-    checkForError()
-  }, [JSON.stringify(values)])
-  const checkForError = () => {
-    let err = false
-    for (const i in values) {
-      if (values[i] == '') {
-        err = true
-        setError((prevState) => ({
-          ...prevState,
-          [i]: 'This field is required',
-        }))
-      } else {
-        setError((prevState) => ({
-          ...prevState,
-          [i]: '',
-        }))
-      }
-    }
-    return err
-  }
-
-  const submitValues = () => {
-    setIsSubmitted(true)
-    if (!checkForError()) {
-      dispatch(checkUserApi(values))
-    }
-    console.log(values)
-  }
-  useEffect(() => {
-    setIsSubmitted(false)
-  }, [isLicenceAPIkeyValid])
 
   const submitLicenseKey = () => {
     setIsSubmitted(true)
@@ -104,77 +60,6 @@ function ApiKey() {
           </div>
         </div>
 
-        {!isLicenceAPIkeyValid ? (
-          <div className="row">
-            <div className="col-md-8 mx-auto">
-              <label>Api Key*</label>
-              <input
-                type="text"
-                className={
-                  submitted && error.apiKey
-                    ? 'border-danger form-control'
-                    : 'form-control'
-                }
-                value={values.apiKey}
-                name="apiKey"
-                onChange={handleChange}
-              />
-              <div className="text-danger">{submitted && error.apiKey}</div>
-              <label>Product Code*</label>
-              <input
-                type="text"
-                className={
-                  submitted && error.productCode
-                    ? 'border-danger form-control'
-                    : 'form-control'
-                }
-                value={values.productCode}
-                name="productCode"
-                onChange={handleChange}
-              />
-              <div className="text-danger">
-                {submitted && error.productCode}
-              </div>
-              <label>Shared Key*</label>
-              <input
-                type="text"
-                className={
-                  submitted && error.sharedKey
-                    ? 'border-danger form-control'
-                    : 'form-control'
-                }
-                value={values.sharedKey}
-                name="sharedKey"
-                onChange={handleChange}
-              />
-              <div className="text-danger">{submitted && error.sharedKey}</div>
-              <label>Management Key*</label>
-              <input
-                type="text"
-                className={
-                  submitted && error.managementKey
-                    ? 'border-danger form-control'
-                    : 'form-control'
-                }
-                value={values.managementKey}
-                name="managementKey"
-                onChange={handleChange}
-              />
-              <div className="text-danger">
-                {submitted && error.managementKey}
-              </div>
-              <button
-                type="button"
-                disabled={isLoading}
-                onClick={() => submitValues()}
-                className="btn mt-3"
-                style={{ backgroundColor: '#00a361', color: 'white' }}
-              >
-                {isLoading ? 'Submitting please wait...' : 'Submit'}
-              </button>
-            </div>
-          </div>
-        ) : (
           <div className="row">
             <div className="col-md-8 mx-auto">
               <label>License Key*</label>
@@ -203,7 +88,6 @@ function ApiKey() {
               </button>
             </div>
           </div>
-        )}
       </div>
     </div>
   )